+++ src/main/java/hu/okrim/productratingapp/repository/RatingRepository.java
@@ -4,5 +4,5 @@ import hu.okrim.productratingapp.entity.Person;
 import hu.okrim.productratingapp.entity.Rating;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RatingRepository extends CrudRepository<Rating, Long> {
+public interface RatingRepository extends CrudRepository<Rating, Integer> {
 }
